cppcheck: fix returnByReference

+++ Source/Lib/Codec/bitstream_unit.h
@@ -368,7 +368,7 @@ static INLINE void aom_write_bit(AomWriter* w, int bit) {
 }
 
 static INLINE void aom_write_literal(AomWriter* w, int data, int bits) {
-    for (int bit = bits - 1; bit >= 0; bit--) aom_write_bit(w, 1 & (data >> bit));
+    for (int bit = bits - 1; bit >= 0; bit--) aom_write_bit(w, 1 & ((unsigned)data >> bit));
 }
 
 static INLINE void aom_write_cdf(AomWriter* w, int symb, const AomCdfProb* cdf, int nsymbs) {
